+++ sas-20181203/src/Models/DeleteHoneypotPresetRequest.php
@@ -19,9 +19,11 @@ class DeleteHoneypotPresetRequest extends Model
     public $honeypotPresetId;
 
     /**
-     * @description 设置请求和接收消息的语言类型，默认为**zh**。取值：
-     * - **zh**：中文
-     * - **en**：英文
+     * @description The language of the content within the request and response. Default value: **zh**. Valid values:
+     *
+     *   **zh**: Chinese.
+     *   **en**: English.
+     *
      * @example zh
      *
      * @var string