+++ sas-20181203/src/Models/DeleteHoneypotProbeRequest.php
@@ -9,10 +9,11 @@ use AlibabaCloud\Tea\Model;
 class DeleteHoneypotProbeRequest extends Model
 {
     /**
-     * @description 设置请求和接收消息的语言类型，默认为**zh**。取值：
+     * @description The language of the content within the request and response. Default value: **zh**. Valid values:
      *
-     * - **zh**：中文
-     * - **en**：英文
+     *   **zh**: Chinese.
+     *   **en**: English.
+     *
      * @example zh
      *
      * @var string