+++ sas-20181203/src/Models/DescribePropertyProcItemResponse.php
@@ -30,9 +30,6 @@ class DescribePropertyProcItemResponse extends Model
 
     public function validate()
     {
-        Model::validateRequired('headers', $this->headers, true);
-        Model::validateRequired('statusCode', $this->statusCode, true);
-        Model::validateRequired('body', $this->body, true);
     }
 
     public function toMap()