+++ sas-20181203/src/Models/ExportWarningRequest.php
@@ -31,7 +31,8 @@ class ExportWarningRequest extends Model
 
     /**
      * @description The ID of the server group.
-     * > You can call the[DescribeAllGroups](~~130972~~)operation to query the IDs of server groups.
+     *
+     * >  You can call the [DescribeAllGroups](~~DescribeAllGroups~~) operation to query the IDs of server groups.
      * @example 13007754
      *
      * @var int